dj - moved enums to proper folder. Updated references in RosterStudent
