+++ src/main/java/edu/ucsb/cs156/example/enums/OrgStatus.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs156.example.services.statuses;
+package edu.ucsb.cs156.example.enums;
 
 public enum OrgStatus {
     NONE, INVITED, MEMBER, OWNER, EXPIRED
