+++ src/main/java/edu/ucsb/cs156/example/enums/RosterStatus.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs156.example.services.statuses;
+package edu.ucsb.cs156.example.enums;
 
 public enum RosterStatus {
     ROSTER, MANUAL, DROPPED
